Fixes #672: adds Java6SoftAssetions to allow use with Android projects
